Rename/split from django_ynh
See discussion here: https://github.com/YunoHost-Apps/django_ynh/pull/13

Corresponding to: https://github.com/YunoHost-Apps/django_example_ynh/pull/1
