--- .github/workflows/package_linter.yml
@@ -1,20 +1,0 @@
-name: YunoHost apps package linter
-
-on:
-  schedule:
-    - cron: '0 8 * * *'
-  push:
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: 'Clone YunoHost apps package linter'
-        run: |
-          git clone --depth=1 https://github.com/YunoHost/package_linter ~/package_linter
-
-      - name: 'Run linter'
-        run: |
-          ~/package_linter/package_linter.py .
