--- conf/django_ynh.service
@@ -1,17 +1,0 @@
-[Unit]
-Description=django_ynh DEMO application server
-After=redis.service postgresql.service
-
-[Service]
-User=__APP__
-Group=__APP__
-WorkingDirectory=__FINALPATH__/
-
-ExecStart=__FINALPATH__/venv/bin/gunicorn --config __FINALPATH__/gunicorn.conf.py wsgi
-
-StandardOutput=syslog
-StandardError=syslog
-SyslogIdentifier=__APP__-server
-
-[Install]
-WantedBy=multi-user.target
