--- conf/wsgi.py
@@ -1,12 +1,0 @@
-"""
-    WSGI config
-"""
-import os
-
-
-os.environ['DJANGO_SETTINGS_MODULE'] = 'settings'
-
-from django.core.wsgi import get_wsgi_application
-
-
-application = get_wsgi_application()
