+++ django_yunohost_integration/management/commands/create_superuser.py
