+++ django_yunohost_integration/secret_key.py
