+++ django_yunohost_integration/views.py
