--- manifest.json
@@ -1,68 +1,0 @@
-{
-	"name": "django_ynh",
-	"id": "django_ynh",
-	"packaging_format": 1,
-	"description": {
-		"en": "Glue code to package django projects as yunohost apps."
-	},
-	"version": "0.1.5~ynh1",
-	"url": "https://github.com/jedie/django_ynh",
-	"license": "GPL-3.0",
-	"maintainer": {
-		"name": "Jens Diemer",
-		"email": "django_ynh@jensdiemer.de"
-	},
-	"previous_maintainers": [],
-	"requirements": {
-		"yunohost": ">= 4.0"
-	},
-	"multi_instance": true,
-	"services": [
-		"nginx"
-	],
-	"arguments": {
-		"install" : [
-			{
-				"name": "domain",
-				"type": "domain",
-				"ask": {
-					"en": "Choose a domain for django_ynh",
-					"fr": "Choisissez un domaine pour django_ynh"
-				},
-				"example": "domain.org"
-			},
-			{
-				"name": "path",
-				"type": "path",
-				"ask": {
-				    "en": "Choose a path for django_ynh",
-				    "fr": "Choisissez un chemin pour django_ynh"
-				},
-				"example": "/django_ynh",
-				"default": "/django_ynh"
-			},
-			{
-				"name": "admin",
-				"type": "user",
-				"ask": {
-				    "en": "Choose an admin user for django_ynh",
-				    "fr": "Choisissez l'administrateur pour django_ynh"
-				},
-				"example": "johndoe"
-			},
-			{
-				"name": "is_public",
-				"type": "boolean",
-				"ask": {
-					"en": "Should django_ynh be public accessible?",
-					"fr": "django_ynh doit-il être accessible au public ?"
-				},
-				"help": {
-					"en": "Any YunoHost user and anonymous people from the web will be able to access the application",
-					"fr": "Tout utilisateur YunoHost et les personnes anonymes pourront accéder à l'application"
-				},
-				"default": false
-			}
-		]
-	}
-}
