--- scripts/backup
@@ -1,69 +1,0 @@
-#!/bin/bash
-
-#=================================================
-# GENERIC START
-#=================================================
-# IMPORT GENERIC HELPERS
-#=================================================
-
-source ../settings/scripts/_common.sh
-source /usr/share/yunohost/helpers
-
-ynh_abort_if_errors
-
-#=================================================
-# LOAD SETTINGS
-#=================================================
-ynh_print_info --message="Loading installation settings..."
-
-app=$YNH_APP_INSTANCE_NAME
-
-public_path=$(ynh_app_setting_get --app="$app" --key=public_path)
-final_path=$(ynh_app_setting_get --app="$app" --key=final_path)
-db_name=$(ynh_app_setting_get --app="$app" --key=db_name)
-
-domain=$(ynh_app_setting_get --app="$app" --key=domain)
-
-#=================================================
-# DECLARE DATA AND CONF FILES TO BACKUP
-#=================================================
-ynh_print_info --message="Declaring files to be backed up..."
-
-#=================================================
-# BACKUP THE APP MAIN DIR
-#=================================================
-
-ynh_backup --src_path="$final_path"
-ynh_backup --src_path="$public_path"
-
-#=================================================
-# BACKUP THE NGINX CONFIGURATION
-#=================================================
-
-ynh_backup --src_path="/etc/nginx/conf.d/$domain.d/$app.conf"
-
-#=================================================
-# BACKUP THE PostgreSQL DATABASE
-#=================================================
-
-ynh_psql_dump_db --database="$db_name" > db.sql
-
-#=================================================
-# SPECIFIC BACKUP
-#=================================================
-# BACKUP LOGROTATE
-#=================================================
-
-ynh_backup --src_path="/etc/logrotate.d/$app"
-
-#=================================================
-# BACKUP SYSTEMD
-#=================================================
-
-ynh_backup --src_path="/etc/systemd/system/$app.service"
-
-#=================================================
-# END OF SCRIPT
-#=================================================
-
-ynh_print_info --message="Backup script completed for $app. (YunoHost will then actually copy those files to the archive)."
