--- scripts/remove
@@ -1,103 +1,0 @@
-#!/bin/bash
-
-#=================================================
-# GENERIC START
-#=================================================
-# IMPORT GENERIC HELPERS
-#=================================================
-
-source _common.sh
-source /usr/share/yunohost/helpers
-
-#=================================================
-# LOAD SETTINGS
-#=================================================
-ynh_script_progression --message="Loading installation settings..."
-
-domain=$(ynh_app_setting_get --app="$app" --key=domain)
-db_name=$(ynh_app_setting_get --app="$app" --key=db_name)
-db_user=$db_name
-public_path=$(ynh_app_setting_get --app="$app" --key=public_path)
-final_path=$(ynh_app_setting_get --app="$app" --key=final_path)
-
-#=================================================
-# STANDARD REMOVE
-#=================================================
-# REMOVE SERVICE FROM ADMIN PANEL
-#=================================================
-
-# Remove a service from the admin panel, added by `yunohost service add`
-if yunohost service status "$app" >/dev/null 2>&1
-then
-	ynh_script_progression --message="Removing $app service integration..."
-	yunohost service remove "$app"
-fi
-
-#=================================================
-# STOP django_ynh'S SERVICES
-#=================================================
-ynh_script_progression --message="Stopping and removing systemd services..." --weight=5
-
-ynh_remove_systemd_config --service="$app"
-
-#=================================================
-# REMOVE THE PostgreSQL DATABASE
-#=================================================
-ynh_script_progression --message="Removing the PostgreSQL database..."
-
-# Remove a database if it exists, along with the associated user
-ynh_psql_remove_db --db_user=$db_user --db_name=$db_name
-
-##=================================================
-## REMOVE REDIS DB
-##=================================================
-
-ynh_redis_remove_db
-
-#=================================================
-# REMOVE DEPENDENCIES
-#=================================================
-ynh_script_progression --message="Removing dependencies..." --weight=10
-
-# Remove metapackage and its dependencies
-ynh_exec_warn_less ynh_remove_app_dependencies
-
-#=================================================
-# REMOVE APP MAIN DIR
-#=================================================
-ynh_script_progression --message="Removing app main directory..."
-
-# Remove the app directory securely
-ynh_secure_remove --file="$public_path"
-ynh_secure_remove --file="$final_path"
-
-#=================================================
-# REMOVE NGINX CONFIGURATION
-#=================================================
-ynh_script_progression --message="Removing nginx web server configuration..."
-
-# Remove the dedicated nginx config
-ynh_remove_nginx_config
-
-#=================================================
-# REMOVE LOGROTATE CONFIGURATION
-#=================================================
-ynh_script_progression --message="Removing logrotate configuration..."
-
-# Remove the app-specific logrotate config
-ynh_remove_logrotate
-
-#=================================================
-# GENERIC FINALIZATION
-#=================================================
-# REMOVE DEDICATED USER
-#=================================================
-ynh_script_progression --message="Removing the dedicated system user..."
-
-# Delete a system user
-ynh_system_user_delete --username="$app"
-
-#=================================================
-# END OF SCRIPT
-#=================================================
-ynh_script_progression --message="Removal of $app completed" --last
